third_party/WebKit/LayoutTests/editing/deleting/delete_keep_style.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
    test(() => assert_selection(
        [
            '<div contenteditable>',
                '<div><b>on^e</b></div><div><i>tw|o</i></div>',
            '</div>',
        ].join(''),
        'delete',
        [
            '<div contenteditable>',
                '<div><b>on|</b><i>o</i></div>',
            '</div>',
        ].join('')),
        'Delete command should keep B and I');
</script>
